+++ Infrastructure/TaskProjectManagement.Persistence/Services/CompanyService.cs
@@ -1,10 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.Design;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using TaskProjectManagement.Application.Interfaces.Services;
 using TaskProjectManagement.Domain.Entities;
+using TaskProjectManagement.Persistence.Errors.NotFoundExceptions;
+
 using TaskProjectManagement.Persistence.Repositories;
 
 namespace TaskProjectManagement.Persistence.Services
@@ -20,6 +23,7 @@ namespace TaskProjectManagement.Persistence.Services
 
         public async Task AddCompanyFromService(Company company)
         {
+
             await _repositoryManager.companyRepository.AddCompany(company); 
             await _repositoryManager.saveChangesAsync();
         }
@@ -27,6 +31,11 @@ namespace TaskProjectManagement.Persistence.Services
         public async Task DeleteCompanyFromService(int companyId)
         {
             var getCompanyForDelete = await _repositoryManager.companyRepository.GetCompanyById(companyId);
+            if (getCompanyForDelete is null)
+            {
+                throw new CompanyNotFoundException(companyId);
+                
+            }
             await _repositoryManager.companyRepository.DeleteCompany(getCompanyForDelete);
             await _repositoryManager.saveChangesAsync();
         }
@@ -38,12 +47,21 @@ namespace TaskProjectManagement.Persistence.Services
 
         public async Task<Company> GetCompanyById(int companyId)
         {
-            return await GetCompanyById(companyId); 
+            var getCompanyForList =  await GetCompanyById(companyId); 
+            if (getCompanyForList is null)
+            {
+                throw new CompanyNotFoundException(companyId);
+            }
+            return getCompanyForList;
         }
 
         public async Task UpdateCompany(Company company)
         {
             var getCompanyForUpdate = await _repositoryManager.companyRepository.GetCompanyById(company.Id);
+            if (getCompanyForUpdate is null)
+            {
+                throw new CompanyNotFoundException(company.Id);
+            }
             getCompanyForUpdate.CompanyName = company.CompanyName;
             getCompanyForUpdate.CompanyLogo = company.CompanyLogo;
             if (company.CompanyMemberCount == 0) {
