+++ Infrastructure/TaskProjectManagement.Persistence/Services/SystemOwnerService.cs
@@ -6,6 +6,8 @@ using System.Threading.Tasks;
 using TaskProjectManagement.Application.Interfaces.Repositories;
 using TaskProjectManagement.Application.Interfaces.Services;
 using TaskProjectManagement.Domain.Entities;
+using TaskProjectManagement.Persistence.Errors.NotFoundExceptions;
+
 
 namespace TaskProjectManagement.Persistence.Services
 {
@@ -20,6 +22,7 @@ namespace TaskProjectManagement.Persistence.Services
 
         public async Task AddOwner(SystemOwner owner)
         {
+
             owner.Password = BCrypt.Net.BCrypt.EnhancedHashPassword(owner.Password);
             await _repositoryManager.systemOwnerRepository.AddOwner(owner);
             await _repositoryManager.saveChangesAsync();
@@ -29,6 +32,10 @@ namespace TaskProjectManagement.Persistence.Services
         public async Task DeleteOwner(int owner)
         {
             var getOwnerForDelete = await _repositoryManager.systemOwnerRepository.GetOwner(owner);
+            if (getOwnerForDelete is null )
+            {
+                throw new SystemOwnerNotFoundException(owner);
+            }
             await _repositoryManager.systemOwnerRepository.DeleteOwner(getOwnerForDelete);
             await _repositoryManager.saveChangesAsync();
         }
@@ -40,12 +47,21 @@ namespace TaskProjectManagement.Persistence.Services
 
         public async Task<SystemOwner> GetOwner(int ownerId)
         {
-            return await _repositoryManager.systemOwnerRepository.GetOwner(ownerId);
+            var getSystemOwner =  await _repositoryManager.systemOwnerRepository.GetOwner(ownerId);
+            if (getSystemOwner is null )
+            {
+                throw new SystemOwnerNotFoundException(ownerId);
+            }
+            return getSystemOwner;
         }
 
         public async Task UpdateOwner(SystemOwner owner)
         {
            var getOwnerForUpdate = await _repositoryManager.systemOwnerRepository.GetOwner(owner.SystemOwnerID);
+            if (getOwnerForUpdate is null)
+            {
+                throw new SystemOwnerNotFoundException(owner.SystemOwnerID);
+            }
            getOwnerForUpdate.Surname = owner.Surname;
            getOwnerForUpdate.Name = owner.Name;
             getOwnerForUpdate.Password = BCrypt.Net.BCrypt.EnhancedHashPassword(owner.Password);
