+++ Infrastructure/TaskProjectManagement.Persistence/Services/TeamLeaderService.cs
@@ -6,6 +6,8 @@ using System.Threading.Tasks;
 using TaskProjectManagement.Application.Interfaces.Repositories;
 using TaskProjectManagement.Application.Interfaces.Services;
 using TaskProjectManagement.Domain.Entities;
+using TaskProjectManagement.Persistence.Errors.NotFoundExceptions;
+
 
 namespace TaskProjectManagement.Persistence.Services
 {
@@ -20,6 +22,7 @@ namespace TaskProjectManagement.Persistence.Services
 
         public async Task AddTeamLeader(TeamLeader teamLeader)
         {
+
             teamLeader.Password = BCrypt.Net.BCrypt.EnhancedHashPassword(teamLeader.Password);
             await _repositoryManager.eamLeaderRepository.AddLeader(teamLeader); 
             await _repositoryManager.saveChangesAsync();
@@ -27,7 +30,12 @@ namespace TaskProjectManagement.Persistence.Services
 
         public async Task<TeamLeader> GetTeamLeader(int teamLeaderId)
         {
-            return await _repositoryManager.eamLeaderRepository.GetLeader(teamLeaderId);
+            var getTeamLeader =  await _repositoryManager.eamLeaderRepository.GetLeader(teamLeaderId);
+            if (getTeamLeader is null)
+            {
+                throw new TeamLeaderNotFoundExceptions(teamLeaderId);
+            }
+            return getTeamLeader;
         }
 
         public async Task<IEnumerable<TeamLeader>> GetTeamLeaderList()
@@ -38,6 +46,10 @@ namespace TaskProjectManagement.Persistence.Services
         public async Task RemoveTeamLeader(int teamLeader)
         {
             var getTeamLeaderForRemove = await _repositoryManager.eamLeaderRepository.GetLeader(teamLeader);
+            if (getTeamLeaderForRemove is null)
+            {
+                throw new TeamLeaderNotFoundExceptions(teamLeader);
+            }
             await _repositoryManager.eamLeaderRepository.RemoveLeader(getTeamLeaderForRemove);
             await _repositoryManager.saveChangesAsync();
 
@@ -46,6 +58,10 @@ namespace TaskProjectManagement.Persistence.Services
         public async Task UpdateTeamLeader(TeamLeader teamLeader)
         {
             var getTeamLeaderForUpdate = await _repositoryManager.eamLeaderRepository.GetLeader(teamLeader.Id);
+            if (getTeamLeaderForUpdate is null)
+            {
+                throw new TeamLeaderNotFoundExceptions(teamLeader.Id);
+            }
             getTeamLeaderForUpdate.Name = teamLeader.Name;
             getTeamLeaderForUpdate.Email = teamLeader.Email;
             getTeamLeaderForUpdate.Password = BCrypt.Net.BCrypt.EnhancedHashPassword(teamLeader.Password);
